Remove duplication in from_shape_ptr and split_at
